Calendar: Show multi-day event continuation
Currently, multi-day events are shown as individual appointments on each day.
This patch ellipsizes multi-day events to indicate continuation on the prior
or following day (or other time-period.)

The time label spot is now replaced by a box layout that contains the prefix
ellipsis label, the time label and the postfix ellipsis label.
In order to keep the alignment, ellipses are merely invisible (zero opacity)
when hidden.

The ellipses are styled using the events-day-time-ellipses class which, by
default, take the color of the event text.

When RTL is used, the box contents are adjusted accordingly (clutter does that
for us).

An event spanning three days now displays "...All Day..." in the calendar on
the second day.

https://bugzilla.gnome.org/show_bug.cgi?id=727302