Calendar: Show multi-day event continuation
Currently, multi-day events are shown as individual appointments on each day.
This patch ellipsizes multi-day events to indicate continuation on the prior
or following day (or other time-period.)

For correct spacing the prefix ellipsis is contained in its own column while
the postfix ellipsis is appended to the time string. See bugzilla for tested
alternatives.

An event spanning three days now displays "...All Day..." in the calendar on
the second day.

https://bugzilla.gnome.org/show_bug.cgi?id=727302