Calendar: sort multi-day events by ending day/time
With commit dc6a60dde, the calendar displays the ending day and time of a
continuing multi-day event on its ending day. This results in the list not
appearing to be sorted. This patch sorts the list according to the displayed
day/time.

With the two appointments
Thursday 0800-1000 Foo, and Wednesday 0900-Friday 1200 Bar and today being
Monday, the rest of the week list currently displays:
F ...1200 Bar
T    0800 Foo
With this patch, the displaying order is switched because Friday comes after
Thursday.

https://bugzilla.gnome.org/show_bug.cgi?id=727302